Test refactored (packages changed, steps defined, etc.)
