+++ src/main/resources/testngLoginTest.xml
@@ -8,7 +8,8 @@
     </groups>
     <test name="Login Page">
         <classes>
-            <class name="com.gl.procamp.tests.functional.smoke.LoginTest"/>
+            <class name="com.gl.procamp.tests.smoke.LoginSmokeTest"/>
+            <class name="com.gl.procamp.tests.api.login.LoginApiTest"/>
         </classes>
     </test>
 </suite>
